Create 15_3Sum.java
the same idea for 2Sum, just needs to use an extra iteration for the third number. Space complexity O(nlogn), because of sorting operation. Time complexity O(n^2). For the rest of the two numbers, using two-pointers. If the sum of those two numbers smaller than the target, choose a larger one from the left pointer, otherwise, choose a smaller one from the right pointer.