Create 56_Merge_Intervals.java
time complexity O(nlogn), space complexity (determined by sorted algorithm). Sorting could simply make the intervals which could be merged into one get together on the index. So we can use two pointers, the one points to the interval we are dealing with(if could, merge it. otherwise, add to result as a new interval), another point to the interval in the result list. (for the convenience that to revise it).