Create 031_Next_Permutation.java
time complexity O(n), space complexity O(1). According to observation, we can find that for every next permutation, the number could be divided into three parts. the first part wouldn't change in next permutation. the second part is one number which would be changed to a larger number chosen from the third part. the third part which is the largest permutation for itself. And after swapping with second part, the third part would be reverse to the smallest permutation for the next permutation. 
So my solution is to find the second part first. then, change the second part. If i==0 means that this permutation is the largest one, we need to reverse it to the smallest. Then, reverse the third part.