Create 289_Game_of_Life.java
time complexity O(n), space complexity O(1). First, update the board with the count of alive neighbors. non-negative numbers mean this cell was dead and the absolute value means the count. positive numbers mean this cell was alive and the value-1 means the count. (because itself also was counted, for being different from the situation that the dead cell without alive neighbor.) Then, according to the new board, we can get the answer by different values.(show in code)