Create 048_Rotate_Image.java
time complexity O(n^2) (the matrix's scale is n*n). space complexity O(1). this is an intuitive solution because we couldn't use extra space. Then we try to move a value one by one and finding out that it would be four times swapping for one iteration. Finally, we could move the value layer by layer.