Create 0206_Reverse_Linked_List.java
Time Complexity O(n), Space complexity O(n). this algorithm is very special. Because we need to reverse it, so from the end of this chain, the next of pointer's next is itself.(head.next.next = head) And we need to delete the original relation of those two nodes.(head.next = null)