Create 0160_Intersection_of_Two_Linked_Lists.java
Time complexity O(m+n), Space complexity O(1). This algorithm is very delicate. List a + List b and List b + List a have the same length. So, the intersection of two singly linked lists begins would be in the tail of those two lists. We could use two-pointer to respectively iterate them and check if they are the same.